clean up project folder layout
Move some package to internal folder
