+++ cmd/containerd-nydus-grpc/snapshotter.go
@@ -5,7 +5,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package snapshotter
+package main
 
 import (
 	"context"
@@ -13,15 +13,43 @@ import (
 	"os"
 	"path/filepath"
 
+	"github.com/pkg/errors"
+
+	"github.com/containerd/nydus-snapshotter/config"
+	"github.com/containerd/nydus-snapshotter/pkg/auth"
+	"github.com/containerd/nydus-snapshotter/pkg/utils/signals"
+	"github.com/containerd/nydus-snapshotter/snapshot"
+
 	api "github.com/containerd/containerd/api/services/snapshots/v1"
 	"github.com/containerd/containerd/contrib/snapshotservice"
 	"github.com/containerd/containerd/log"
 	"github.com/containerd/containerd/snapshots"
-	"github.com/pkg/errors"
 	"google.golang.org/grpc"
-
-	"github.com/containerd/nydus-snapshotter/pkg/auth"
 )
+
+func Start(ctx context.Context, cfg *config.SnapshotterConfig) error {
+	ctx, cancel := context.WithCancel(ctx)
+	defer cancel()
+	rs, err := snapshot.NewSnapshotter(ctx, cfg)
+	if err != nil {
+		return errors.Wrap(err, "failed to initialize snapshotter")
+	}
+
+	stopSignal := signals.SetupSignalHandler()
+	opt := ServeOptions{
+		ListeningSocketPath: cfg.Address,
+		EnableCRIKeychain:   cfg.RemoteConfig.AuthConfig.EnableCRIKeychain,
+		ImageServiceAddress: cfg.RemoteConfig.AuthConfig.ImageServiceAddress,
+	}
+
+	if cfg.RemoteConfig.AuthConfig.EnableKubeconfigKeychain {
+		if err := auth.InitKubeSecretListener(ctx, cfg.RemoteConfig.AuthConfig.KubeconfigPath); err != nil {
+			return err
+		}
+	}
+
+	return Serve(ctx, rs, opt, stopSignal)
+}
 
 type ServeOptions struct {
 	ListeningSocketPath string
