+++ internal/flags/flags_test.go
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package command
+package flags
 
 import (
 	"flag"
